+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - `snakemake` to 8.14
   - `neutcurve` to 2.1
   - `papermill` to 2.6
+- Added test that compares actual results to expected ones for `test_example` better testing of pipeline updates.
 
 #### version 3.14.1
 - Make change in version 3.14.0 backward compatible by assuming `use_precomputed_barcode_counts` is `False` if not specified.